% Generated by build script
\name{olib_book}
\alias{olib_book}
\title{Get book details by ISBN from Open Library}
\usage{
olib_book()
}
\arguments{
\item{isbn}{ISBN-10 or ISBN-13 (e.g. "9780140328721", "0140328726")}
}
\value{tibble: title, authors, publishers, publish_date,}
\description{Get book details by ISBN from Open Library}
