% Generated by build script
\name{olib_subject}
\alias{olib_subject}
\title{Search Open Library by subject}
\usage{
olib_subject()
}
\arguments{
\item{subject}{Subject name (e.g. "science fiction", "history")}
\item{limit}{Number of results (default 20)}
}
\value{tibble: key, title, author_name, first_publish_year}
\description{Search Open Library by subject}
