% Generated by build script
\name{olib_search}
\alias{olib_search}
\title{Search Open Library for books}
\usage{
olib_search()
}
\arguments{
\item{query}{Search query (e.g. "lord of the rings", "machine learning")}
\item{limit}{Number of results (default 20, max 100)}
}
\value{tibble: key, title, author_name, first_publish_year, publisher,}
\description{Search Open Library for books}
